Revert "Rename file from upper camel to kebab case"
This reverts commit d62e6c56ba10d3081ca9fe7a18ed3db1e5608429.
