+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from "@/components/footer";
-import { Header } from "@/components/header";
+import { Footer } from "@/components/Footer";
+import { Header } from "@/components/Header";
 import "@/styles/globals.css";
 import { Metadata } from "next";
 import { ReactNode } from "react";
