+++ src/app/skills/page.tsx
@@ -1,5 +1,5 @@
 import { TechnicalSkills } from "@/app/skills/components/TechnicalSkills";
-import { PageTitle } from "@/components/page-title";
+import { PageTitle } from "@/components/PageTitle";
 import { Metadata } from "next";
 
 const title = "技術スタック - ryo-manba";
