+++ src/components/Footer.stories.tsx
@@ -1,11 +1,11 @@
-import type { StoryObj } from "@storybook/react";
+import type { StoryObj } from '@storybook/react';
 
-import { Footer } from "./footer";
+import { Footer } from './Footer';
 
 const meta = {
-  title: "Footer",
+  title: 'Footer',
   component: Footer,
-};
+} 
 
 export default meta;
 type Story = StoryObj<typeof meta>;
