+++ src/components/Header.stories.tsx
@@ -1,11 +1,11 @@
-import type { StoryObj } from "@storybook/react";
+import type { StoryObj } from '@storybook/react';
 
-import { Header } from "./header";
+import { Header } from './Header';
 
 const meta = {
-  title: "Header",
+  title: 'Header',
   component: Header,
-};
+} 
 
 export default meta;
 type Story = StoryObj<typeof meta>;
